refactor: keep code in a single file, add license reference
