+++ qaul_ui/lib/screens/home/dynamic_network/dynamic_network_screen.dart
@@ -14,7 +14,6 @@ import 'package:flutter/material.dart' hide Draggable;
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:hooks_riverpod/legacy.dart';
-import 'package:open_simplex_2/open_simplex_2.dart';
 import 'package:qaul_rpc/qaul_rpc.dart';
 import 'package:utils/utils.dart';
 
@@ -168,7 +167,7 @@ class _NetworkNodeComponent extends BodyComponent
 
   final _NetworkNodeComponent? ballParent;
 
-  final _noise = OpenSimplex2F(math.Random().nextInt(255));
+  final _noise = OpenSimplexNoise(math.Random().nextInt(255));
 
   async.Timer? _timer;
