--- qaul_ui/packages/utils/lib/src/grad.dart
@@ -1,17 +1,0 @@
-/// 4-dimensional gradient.
-class Grad4 {
-  /// Creates a 4-dimensional gradient from its components.
-  const Grad4(this.dx, this.dy, this.dz, this.dw);
-
-  /// The x component of the gradient.
-  final double dx;
-
-  /// The y component of the gradient.
-  final double dy;
-
-  /// The z component of the gradient.
-  final double dz;
-
-  /// The w component of the gradinet.
-  final double dw;
-}
