+++ qaul_ui/packages/utils/lib/src/noise.dart
@@ -1,0 +1,1263 @@
+// From https://github.com/creativecreatorormaybenot/funvas/tree/main/open_simplex_2
+// Copyright (c) 2021-2025, creativecreatorormaybenot
+// Licensed under BSD-3-Clause. See docs/licenses/THIRD_PARTY_LICENSES.md
+
+// The implementation in this file is based on KdotJPG's implementation here: https://github.com/KdotJPG/OpenSimplex2/blob/a186b9bb644747c936d7cba748d11f28b1cee66e/java/OpenSimplex2F.java.
+
+import 'dart:typed_data';
+
+import 'package:fixnum/fixnum.dart';
+
+/// 4-dimensional gradient.
+class _Grad4 {
+  /// Creates a 4-dimensional gradient from its components.
+  const _Grad4(this.dx, this.dy, this.dz, this.dw);
+
+  /// The x component of the gradient.
+  final double dx;
+
+  /// The y component of the gradient.
+  final double dy;
+
+  /// The z component of the gradient.
+  final double dz;
+
+  /// The w component of the gradinet.
+  final double dw;
+}
+
+const _kN4 = 0.009202377986303158;
+const _kSize = 2048;
+const _kMask = 2047;
+
+/// K.jpg's OpenSimplex 2, faster variant.
+///
+/// - 2D is standard simplex implemented using a lookup table.
+/// - 3D is "Re-oriented 4-point BCC noise" which constructs a
+///   congruent BCC lattice in a much different way than usual.
+/// - 4D constructs the lattice as a union of five copies of its
+///   reciprocal. It successively finds the closest point on each.
+///
+/// Multiple versions of each function are provided. See the documentation for
+/// each for more info.
+class OpenSimplexNoise {
+  /// Creates a seeded [OpenSimplexNoise] that can be used to evaluate noise.
+  OpenSimplexNoise(int seed) {
+    if (!_staticInitialized) {
+      _staticInit();
+      _staticInitialized = true;
+    }
+
+    final source = Int16List(_kSize);
+    for (var i = 0; i < _kSize; i++) {
+      source[i] = i;
+    }
+    // KdotJPG's implementation uses Java's long here. Long in Java is a
+    // 64-bit two's complement integer. int in Dart is also a 64-bit two's
+    // complement integer, however, *only on native* (see https://dart.dev/guides/language/numbers).
+    // However, we want to support web, i.e. JavaScript, as well with this
+    // package and therefore we have to use the fixnum Int64 type. See
+    // https://github.com/dart-lang/sdk/issues/46852#issuecomment-894888740.
+    var seed64 = Int64(seed);
+    for (int i = _kSize - 1; i >= 0; i--) {
+      // KdotJPG's implementation uses long literals here. We can use int
+      // literals of this size as well in Dart, however, these are too big for
+      // JavaScript and therefore we have to use int.parse instead.
+      seed64 =
+          seed64 * Int64.parseInt('6364136223846793005') +
+          Int64.parseInt('1442695040888963407');
+      // We know r cannot be bigger than 2047, so we can convert it back to an
+      // int.
+      var r = ((seed64 + 31) % (i + 1)).toInt();
+      if (r < 0) r += i + 1;
+      _perm[i] = source[r];
+      _permGrad4[i] = _gradients4d[_perm[i]];
+      source[r] = source[i];
+    }
+  }
+
+  final _perm = Int16List(_kSize);
+  final _permGrad4 = List.filled(_kSize, const _Grad4(0, 0, 0, 0));
+
+  /// 4D OpenSimplex2F noise, classic lattice orientation.
+  double noise4Classic(double x, double y, double z, double w) {
+    // Get points for A4 lattice
+    double s = -0.138196601125011 * (x + y + z + w);
+    double xs = x + s, ys = y + s, zs = z + s, ws = w + s;
+
+    return _noise4Base(xs, ys, zs, ws);
+  }
+
+  /// 4D OpenSimplex2F noise base.
+  ///
+  /// Current implementation not fully optimized by lookup tables.
+  /// But still comes out slightly ahead of Gustavson's Simplex in tests.
+  double _noise4Base(double xs, double ys, double zs, double ws) {
+    double value = 0;
+
+    // Get base points and offsets
+    int xsb = xs.floor(), ysb = ys.floor(), zsb = zs.floor(), wsb = ws.floor();
+    double xsi = xs - xsb, ysi = ys - ysb, zsi = zs - zsb, wsi = ws - wsb;
+
+    // If we're in the lower half, flip so we can repeat the code for the upper half. We'll flip back later.
+    double siSum = xsi + ysi + zsi + wsi;
+    double ssi = siSum * 0.309016994374947; // Prep for vertex contributions.
+    final inLowerHalf = (siSum < 2);
+    if (inLowerHalf) {
+      xsi = 1 - xsi;
+      ysi = 1 - ysi;
+      zsi = 1 - zsi;
+      wsi = 1 - wsi;
+      siSum = 4 - siSum;
+    }
+
+    // Consider opposing vertex pairs of the octahedron formed by the central cross-section of the stretched tesseract
+    double aabb = xsi + ysi - zsi - wsi,
+        abab = xsi - ysi + zsi - wsi,
+        abba = xsi - ysi - zsi + wsi;
+    double aabbScore = aabb.abs(),
+        ababScore = abab.abs(),
+        abbaScore = abba.abs();
+
+    // Find the closest point on the stretched tesseract as if it were the upper half
+    int vertexIndex, via, vib;
+    double asi, bsi;
+    if (aabbScore > ababScore && aabbScore > abbaScore) {
+      if (aabb > 0) {
+        asi = zsi;
+        bsi = wsi;
+        vertexIndex = int.parse('0011', radix: 2);
+        via = int.parse('0111', radix: 2);
+        vib = int.parse('1011', radix: 2);
+      } else {
+        asi = xsi;
+        bsi = ysi;
+        vertexIndex = int.parse('1100', radix: 2);
+        via = int.parse('1101', radix: 2);
+        vib = int.parse('1110', radix: 2);
+      }
+    } else if (ababScore > abbaScore) {
+      if (abab > 0) {
+        asi = ysi;
+        bsi = wsi;
+        vertexIndex = int.parse('0101', radix: 2);
+        via = int.parse('0111', radix: 2);
+        vib = int.parse('1101', radix: 2);
+      } else {
+        asi = xsi;
+        bsi = zsi;
+        vertexIndex = int.parse('1010', radix: 2);
+        via = int.parse('1011', radix: 2);
+        vib = int.parse('1110', radix: 2);
+      }
+    } else {
+      if (abba > 0) {
+        asi = ysi;
+        bsi = zsi;
+        vertexIndex = int.parse('1001', radix: 2);
+        via = int.parse('1011', radix: 2);
+        vib = int.parse('1101', radix: 2);
+      } else {
+        asi = xsi;
+        bsi = wsi;
+        vertexIndex = int.parse('0110', radix: 2);
+        via = int.parse('0111', radix: 2);
+        vib = int.parse('1110', radix: 2);
+      }
+    }
+    if (bsi > asi) {
+      via = vib;
+      double temp = bsi;
+      bsi = asi;
+      asi = temp;
+    }
+    if (siSum + asi > 3) {
+      vertexIndex = via;
+      if (siSum + bsi > 4) {
+        vertexIndex = int.parse('1111', radix: 2);
+      }
+    }
+
+    // Now flip back if we're actually in the lower half.
+    if (inLowerHalf) {
+      xsi = 1 - xsi;
+      ysi = 1 - ysi;
+      zsi = 1 - zsi;
+      wsi = 1 - wsi;
+      vertexIndex ^= int.parse('1111', radix: 2);
+    }
+
+    // Five points to add, total, from five copies of the A4 lattice.
+    for (int i = 0; i < 5; i++) {
+      // Update xsb/etc. and add the lattice point's contribution.
+      _LatticePoint4D c = _vertices4d[vertexIndex];
+      xsb += c.xsv;
+      ysb += c.ysv;
+      zsb += c.zsv;
+      wsb += c.wsv;
+      double xi = xsi + ssi, yi = ysi + ssi, zi = zsi + ssi, wi = wsi + ssi;
+      double dx = xi + c.dx, dy = yi + c.dy, dz = zi + c.dz, dw = wi + c.dw;
+      double attn = 0.5 - dx * dx - dy * dy - dz * dz - dw * dw;
+      if (attn > 0) {
+        int pxm = xsb & _kMask,
+            pym = ysb & _kMask,
+            pzm = zsb & _kMask,
+            pwm = wsb & _kMask;
+        _Grad4 grad = _permGrad4[_perm[_perm[_perm[pxm] ^ pym] ^ pzm] ^ pwm];
+        double ramped =
+            grad.dx * dx + grad.dy * dy + grad.dz * dz + grad.dw * dw;
+
+        attn *= attn;
+        value += attn * attn * ramped;
+      }
+
+      // Maybe this helps the compiler/JVM/LLVM/etc. know we can end the loop here. Maybe not.
+      if (i == 4) break;
+
+      // Update the relative skewed coordinates to reference the vertex we just added.
+      // Rather, reference its counterpart on the lattice copy that is shifted down by
+      // the vector <-0.2, -0.2, -0.2, -0.2>
+      xsi += c.xsi;
+      ysi += c.ysi;
+      zsi += c.zsi;
+      wsi += c.wsi;
+      ssi += c.ssiDelta;
+
+      // Next point is the closest vertex on the 4-simplex whose base vertex is the aforementioned vertex.
+      double score0 =
+          1.0 +
+          ssi *
+              (-1.0 /
+                  0.309016994374947); // Seems slightly faster than 1.0-xsi-ysi-zsi-wsi
+      vertexIndex = int.parse('0000', radix: 2);
+      if (xsi >= ysi && xsi >= zsi && xsi >= wsi && xsi >= score0) {
+        vertexIndex = int.parse('0001', radix: 2);
+      } else if (ysi > xsi && ysi >= zsi && ysi >= wsi && ysi >= score0) {
+        vertexIndex = int.parse('0010', radix: 2);
+      } else if (zsi > xsi && zsi > ysi && zsi >= wsi && zsi >= score0) {
+        vertexIndex = int.parse('0100', radix: 2);
+      } else if (wsi > xsi && wsi > ysi && wsi > zsi && wsi >= score0) {
+        vertexIndex = int.parse('1000', radix: 2);
+      }
+    }
+
+    return value;
+  }
+
+  // Definitions
+
+  static final _vertices4d = List.filled(16, _LatticePoint4D(0, 0, 0, 0));
+
+  static final _gradients4d = <_Grad4>[];
+
+  static var _staticInitialized = false;
+
+  /// Performs the initialization of all static lookup members.
+  ///
+  /// This function as well as [_staticInitialized] exist because there is
+  /// no comparable concept to static blocks (from Java) in Dart.
+  static void _staticInit() {
+    for (int i = 0; i < 16; i++) {
+      _vertices4d[i] = _LatticePoint4D(
+        (i >> 0) & 1,
+        (i >> 1) & 1,
+        (i >> 2) & 1,
+        (i >> 3) & 1,
+      );
+    }
+
+    const grad4 = [
+      _Grad4(
+        -0.753341017856078,
+        -0.37968289875261624,
+        -0.37968289875261624,
+        -0.37968289875261624,
+      ),
+      _Grad4(
+        -0.7821684431180708,
+        -0.4321472685365301,
+        -0.4321472685365301,
+        0.12128480194602098,
+      ),
+      _Grad4(
+        -0.7821684431180708,
+        -0.4321472685365301,
+        0.12128480194602098,
+        -0.4321472685365301,
+      ),
+      _Grad4(
+        -0.7821684431180708,
+        0.12128480194602098,
+        -0.4321472685365301,
+        -0.4321472685365301,
+      ),
+      _Grad4(
+        -0.8586508742123365,
+        -0.508629699630796,
+        0.044802370851755174,
+        0.044802370851755174,
+      ),
+      _Grad4(
+        -0.8586508742123365,
+        0.044802370851755174,
+        -0.508629699630796,
+        0.044802370851755174,
+      ),
+      _Grad4(
+        -0.8586508742123365,
+        0.044802370851755174,
+        0.044802370851755174,
+        -0.508629699630796,
+      ),
+      _Grad4(
+        -0.9982828964265062,
+        -0.03381941603233842,
+        -0.03381941603233842,
+        -0.03381941603233842,
+      ),
+      _Grad4(
+        -0.37968289875261624,
+        -0.753341017856078,
+        -0.37968289875261624,
+        -0.37968289875261624,
+      ),
+      _Grad4(
+        -0.4321472685365301,
+        -0.7821684431180708,
+        -0.4321472685365301,
+        0.12128480194602098,
+      ),
+      _Grad4(
+        -0.4321472685365301,
+        -0.7821684431180708,
+        0.12128480194602098,
+        -0.4321472685365301,
+      ),
+      _Grad4(
+        0.12128480194602098,
+        -0.7821684431180708,
+        -0.4321472685365301,
+        -0.4321472685365301,
+      ),
+      _Grad4(
+        -0.508629699630796,
+        -0.8586508742123365,
+        0.044802370851755174,
+        0.044802370851755174,
+      ),
+      _Grad4(
+        0.044802370851755174,
+        -0.8586508742123365,
+        -0.508629699630796,
+        0.044802370851755174,
+      ),
+      _Grad4(
+        0.044802370851755174,
+        -0.8586508742123365,
+        0.044802370851755174,
+        -0.508629699630796,
+      ),
+      _Grad4(
+        -0.03381941603233842,
+        -0.9982828964265062,
+        -0.03381941603233842,
+        -0.03381941603233842,
+      ),
+      _Grad4(
+        -0.37968289875261624,
+        -0.37968289875261624,
+        -0.753341017856078,
+        -0.37968289875261624,
+      ),
+      _Grad4(
+        -0.4321472685365301,
+        -0.4321472685365301,
+        -0.7821684431180708,
+        0.12128480194602098,
+      ),
+      _Grad4(
+        -0.4321472685365301,
+        0.12128480194602098,
+        -0.7821684431180708,
+        -0.4321472685365301,
+      ),
+      _Grad4(
+        0.12128480194602098,
+        -0.4321472685365301,
+        -0.7821684431180708,
+        -0.4321472685365301,
+      ),
+      _Grad4(
+        -0.508629699630796,
+        0.044802370851755174,
+        -0.8586508742123365,
+        0.044802370851755174,
+      ),
+      _Grad4(
+        0.044802370851755174,
+        -0.508629699630796,
+        -0.8586508742123365,
+        0.044802370851755174,
+      ),
+      _Grad4(
+        0.044802370851755174,
+        0.044802370851755174,
+        -0.8586508742123365,
+        -0.508629699630796,
+      ),
+      _Grad4(
+        -0.03381941603233842,
+        -0.03381941603233842,
+        -0.9982828964265062,
+        -0.03381941603233842,
+      ),
+      _Grad4(
+        -0.37968289875261624,
+        -0.37968289875261624,
+        -0.37968289875261624,
+        -0.753341017856078,
+      ),
+      _Grad4(
+        -0.4321472685365301,
+        -0.4321472685365301,
+        0.12128480194602098,
+        -0.7821684431180708,
+      ),
+      _Grad4(
+        -0.4321472685365301,
+        0.12128480194602098,
+        -0.4321472685365301,
+        -0.7821684431180708,
+      ),
+      _Grad4(
+        0.12128480194602098,
+        -0.4321472685365301,
+        -0.4321472685365301,
+        -0.7821684431180708,
+      ),
+      _Grad4(
+        -0.508629699630796,
+        0.044802370851755174,
+        0.044802370851755174,
+        -0.8586508742123365,
+      ),
+      _Grad4(
+        0.044802370851755174,
+        -0.508629699630796,
+        0.044802370851755174,
+        -0.8586508742123365,
+      ),
+      _Grad4(
+        0.044802370851755174,
+        0.044802370851755174,
+        -0.508629699630796,
+        -0.8586508742123365,
+      ),
+      _Grad4(
+        -0.03381941603233842,
+        -0.03381941603233842,
+        -0.03381941603233842,
+        -0.9982828964265062,
+      ),
+      _Grad4(
+        -0.6740059517812944,
+        -0.3239847771997537,
+        -0.3239847771997537,
+        0.5794684678643381,
+      ),
+      _Grad4(
+        -0.7504883828755602,
+        -0.4004672082940195,
+        0.15296486218853164,
+        0.5029860367700724,
+      ),
+      _Grad4(
+        -0.7504883828755602,
+        0.15296486218853164,
+        -0.4004672082940195,
+        0.5029860367700724,
+      ),
+      _Grad4(
+        -0.8828161875373585,
+        0.08164729285680945,
+        0.08164729285680945,
+        0.4553054119602712,
+      ),
+      _Grad4(
+        -0.4553054119602712,
+        -0.08164729285680945,
+        -0.08164729285680945,
+        0.8828161875373585,
+      ),
+      _Grad4(
+        -0.5029860367700724,
+        -0.15296486218853164,
+        0.4004672082940195,
+        0.7504883828755602,
+      ),
+      _Grad4(
+        -0.5029860367700724,
+        0.4004672082940195,
+        -0.15296486218853164,
+        0.7504883828755602,
+      ),
+      _Grad4(
+        -0.5794684678643381,
+        0.3239847771997537,
+        0.3239847771997537,
+        0.6740059517812944,
+      ),
+      _Grad4(
+        -0.3239847771997537,
+        -0.6740059517812944,
+        -0.3239847771997537,
+        0.5794684678643381,
+      ),
+      _Grad4(
+        -0.4004672082940195,
+        -0.7504883828755602,
+        0.15296486218853164,
+        0.5029860367700724,
+      ),
+      _Grad4(
+        0.15296486218853164,
+        -0.7504883828755602,
+        -0.4004672082940195,
+        0.5029860367700724,
+      ),
+      _Grad4(
+        0.08164729285680945,
+        -0.8828161875373585,
+        0.08164729285680945,
+        0.4553054119602712,
+      ),
+      _Grad4(
+        -0.08164729285680945,
+        -0.4553054119602712,
+        -0.08164729285680945,
+        0.8828161875373585,
+      ),
+      _Grad4(
+        -0.15296486218853164,
+        -0.5029860367700724,
+        0.4004672082940195,
+        0.7504883828755602,
+      ),
+      _Grad4(
+        0.4004672082940195,
+        -0.5029860367700724,
+        -0.15296486218853164,
+        0.7504883828755602,
+      ),
+      _Grad4(
+        0.3239847771997537,
+        -0.5794684678643381,
+        0.3239847771997537,
+        0.6740059517812944,
+      ),
+      _Grad4(
+        -0.3239847771997537,
+        -0.3239847771997537,
+        -0.6740059517812944,
+        0.5794684678643381,
+      ),
+      _Grad4(
+        -0.4004672082940195,
+        0.15296486218853164,
+        -0.7504883828755602,
+        0.5029860367700724,
+      ),
+      _Grad4(
+        0.15296486218853164,
+        -0.4004672082940195,
+        -0.7504883828755602,
+        0.5029860367700724,
+      ),
+      _Grad4(
+        0.08164729285680945,
+        0.08164729285680945,
+        -0.8828161875373585,
+        0.4553054119602712,
+      ),
+      _Grad4(
+        -0.08164729285680945,
+        -0.08164729285680945,
+        -0.4553054119602712,
+        0.8828161875373585,
+      ),
+      _Grad4(
+        -0.15296486218853164,
+        0.4004672082940195,
+        -0.5029860367700724,
+        0.7504883828755602,
+      ),
+      _Grad4(
+        0.4004672082940195,
+        -0.15296486218853164,
+        -0.5029860367700724,
+        0.7504883828755602,
+      ),
+      _Grad4(
+        0.3239847771997537,
+        0.3239847771997537,
+        -0.5794684678643381,
+        0.6740059517812944,
+      ),
+      _Grad4(
+        -0.6740059517812944,
+        -0.3239847771997537,
+        0.5794684678643381,
+        -0.3239847771997537,
+      ),
+      _Grad4(
+        -0.7504883828755602,
+        -0.4004672082940195,
+        0.5029860367700724,
+        0.15296486218853164,
+      ),
+      _Grad4(
+        -0.7504883828755602,
+        0.15296486218853164,
+        0.5029860367700724,
+        -0.4004672082940195,
+      ),
+      _Grad4(
+        -0.8828161875373585,
+        0.08164729285680945,
+        0.4553054119602712,
+        0.08164729285680945,
+      ),
+      _Grad4(
+        -0.4553054119602712,
+        -0.08164729285680945,
+        0.8828161875373585,
+        -0.08164729285680945,
+      ),
+      _Grad4(
+        -0.5029860367700724,
+        -0.15296486218853164,
+        0.7504883828755602,
+        0.4004672082940195,
+      ),
+      _Grad4(
+        -0.5029860367700724,
+        0.4004672082940195,
+        0.7504883828755602,
+        -0.15296486218853164,
+      ),
+      _Grad4(
+        -0.5794684678643381,
+        0.3239847771997537,
+        0.6740059517812944,
+        0.3239847771997537,
+      ),
+      _Grad4(
+        -0.3239847771997537,
+        -0.6740059517812944,
+        0.5794684678643381,
+        -0.3239847771997537,
+      ),
+      _Grad4(
+        -0.4004672082940195,
+        -0.7504883828755602,
+        0.5029860367700724,
+        0.15296486218853164,
+      ),
+      _Grad4(
+        0.15296486218853164,
+        -0.7504883828755602,
+        0.5029860367700724,
+        -0.4004672082940195,
+      ),
+      _Grad4(
+        0.08164729285680945,
+        -0.8828161875373585,
+        0.4553054119602712,
+        0.08164729285680945,
+      ),
+      _Grad4(
+        -0.08164729285680945,
+        -0.4553054119602712,
+        0.8828161875373585,
+        -0.08164729285680945,
+      ),
+      _Grad4(
+        -0.15296486218853164,
+        -0.5029860367700724,
+        0.7504883828755602,
+        0.4004672082940195,
+      ),
+      _Grad4(
+        0.4004672082940195,
+        -0.5029860367700724,
+        0.7504883828755602,
+        -0.15296486218853164,
+      ),
+      _Grad4(
+        0.3239847771997537,
+        -0.5794684678643381,
+        0.6740059517812944,
+        0.3239847771997537,
+      ),
+      _Grad4(
+        -0.3239847771997537,
+        -0.3239847771997537,
+        0.5794684678643381,
+        -0.6740059517812944,
+      ),
+      _Grad4(
+        -0.4004672082940195,
+        0.15296486218853164,
+        0.5029860367700724,
+        -0.7504883828755602,
+      ),
+      _Grad4(
+        0.15296486218853164,
+        -0.4004672082940195,
+        0.5029860367700724,
+        -0.7504883828755602,
+      ),
+      _Grad4(
+        0.08164729285680945,
+        0.08164729285680945,
+        0.4553054119602712,
+        -0.8828161875373585,
+      ),
+      _Grad4(
+        -0.08164729285680945,
+        -0.08164729285680945,
+        0.8828161875373585,
+        -0.4553054119602712,
+      ),
+      _Grad4(
+        -0.15296486218853164,
+        0.4004672082940195,
+        0.7504883828755602,
+        -0.5029860367700724,
+      ),
+      _Grad4(
+        0.4004672082940195,
+        -0.15296486218853164,
+        0.7504883828755602,
+        -0.5029860367700724,
+      ),
+      _Grad4(
+        0.3239847771997537,
+        0.3239847771997537,
+        0.6740059517812944,
+        -0.5794684678643381,
+      ),
+      _Grad4(
+        -0.6740059517812944,
+        0.5794684678643381,
+        -0.3239847771997537,
+        -0.3239847771997537,
+      ),
+      _Grad4(
+        -0.7504883828755602,
+        0.5029860367700724,
+        -0.4004672082940195,
+        0.15296486218853164,
+      ),
+      _Grad4(
+        -0.7504883828755602,
+        0.5029860367700724,
+        0.15296486218853164,
+        -0.4004672082940195,
+      ),
+      _Grad4(
+        -0.8828161875373585,
+        0.4553054119602712,
+        0.08164729285680945,
+        0.08164729285680945,
+      ),
+      _Grad4(
+        -0.4553054119602712,
+        0.8828161875373585,
+        -0.08164729285680945,
+        -0.08164729285680945,
+      ),
+      _Grad4(
+        -0.5029860367700724,
+        0.7504883828755602,
+        -0.15296486218853164,
+        0.4004672082940195,
+      ),
+      _Grad4(
+        -0.5029860367700724,
+        0.7504883828755602,
+        0.4004672082940195,
+        -0.15296486218853164,
+      ),
+      _Grad4(
+        -0.5794684678643381,
+        0.6740059517812944,
+        0.3239847771997537,
+        0.3239847771997537,
+      ),
+      _Grad4(
+        -0.3239847771997537,
+        0.5794684678643381,
+        -0.6740059517812944,
+        -0.3239847771997537,
+      ),
+      _Grad4(
+        -0.4004672082940195,
+        0.5029860367700724,
+        -0.7504883828755602,
+        0.15296486218853164,
+      ),
+      _Grad4(
+        0.15296486218853164,
+        0.5029860367700724,
+        -0.7504883828755602,
+        -0.4004672082940195,
+      ),
+      _Grad4(
+        0.08164729285680945,
+        0.4553054119602712,
+        -0.8828161875373585,
+        0.08164729285680945,
+      ),
+      _Grad4(
+        -0.08164729285680945,
+        0.8828161875373585,
+        -0.4553054119602712,
+        -0.08164729285680945,
+      ),
+      _Grad4(
+        -0.15296486218853164,
+        0.7504883828755602,
+        -0.5029860367700724,
+        0.4004672082940195,
+      ),
+      _Grad4(
+        0.4004672082940195,
+        0.7504883828755602,
+        -0.5029860367700724,
+        -0.15296486218853164,
+      ),
+      _Grad4(
+        0.3239847771997537,
+        0.6740059517812944,
+        -0.5794684678643381,
+        0.3239847771997537,
+      ),
+      _Grad4(
+        -0.3239847771997537,
+        0.5794684678643381,
+        -0.3239847771997537,
+        -0.6740059517812944,
+      ),
+      _Grad4(
+        -0.4004672082940195,
+        0.5029860367700724,
+        0.15296486218853164,
+        -0.7504883828755602,
+      ),
+      _Grad4(
+        0.15296486218853164,
+        0.5029860367700724,
+        -0.4004672082940195,
+        -0.7504883828755602,
+      ),
+      _Grad4(
+        0.08164729285680945,
+        0.4553054119602712,
+        0.08164729285680945,
+        -0.8828161875373585,
+      ),
+      _Grad4(
+        -0.08164729285680945,
+        0.8828161875373585,
+        -0.08164729285680945,
+        -0.4553054119602712,
+      ),
+      _Grad4(
+        -0.15296486218853164,
+        0.7504883828755602,
+        0.4004672082940195,
+        -0.5029860367700724,
+      ),
+      _Grad4(
+        0.4004672082940195,
+        0.7504883828755602,
+        -0.15296486218853164,
+        -0.5029860367700724,
+      ),
+      _Grad4(
+        0.3239847771997537,
+        0.6740059517812944,
+        0.3239847771997537,
+        -0.5794684678643381,
+      ),
+      _Grad4(
+        0.5794684678643381,
+        -0.6740059517812944,
+        -0.3239847771997537,
+        -0.3239847771997537,
+      ),
+      _Grad4(
+        0.5029860367700724,
+        -0.7504883828755602,
+        -0.4004672082940195,
+        0.15296486218853164,
+      ),
+      _Grad4(
+        0.5029860367700724,
+        -0.7504883828755602,
+        0.15296486218853164,
+        -0.4004672082940195,
+      ),
+      _Grad4(
+        0.4553054119602712,
+        -0.8828161875373585,
+        0.08164729285680945,
+        0.08164729285680945,
+      ),
+      _Grad4(
+        0.8828161875373585,
+        -0.4553054119602712,
+        -0.08164729285680945,
+        -0.08164729285680945,
+      ),
+      _Grad4(
+        0.7504883828755602,
+        -0.5029860367700724,
+        -0.15296486218853164,
+        0.4004672082940195,
+      ),
+      _Grad4(
+        0.7504883828755602,
+        -0.5029860367700724,
+        0.4004672082940195,
+        -0.15296486218853164,
+      ),
+      _Grad4(
+        0.6740059517812944,
+        -0.5794684678643381,
+        0.3239847771997537,
+        0.3239847771997537,
+      ),
+      _Grad4(
+        0.5794684678643381,
+        -0.3239847771997537,
+        -0.6740059517812944,
+        -0.3239847771997537,
+      ),
+      _Grad4(
+        0.5029860367700724,
+        -0.4004672082940195,
+        -0.7504883828755602,
+        0.15296486218853164,
+      ),
+      _Grad4(
+        0.5029860367700724,
+        0.15296486218853164,
+        -0.7504883828755602,
+        -0.4004672082940195,
+      ),
+      _Grad4(
+        0.4553054119602712,
+        0.08164729285680945,
+        -0.8828161875373585,
+        0.08164729285680945,
+      ),
+      _Grad4(
+        0.8828161875373585,
+        -0.08164729285680945,
+        -0.4553054119602712,
+        -0.08164729285680945,
+      ),
+      _Grad4(
+        0.7504883828755602,
+        -0.15296486218853164,
+        -0.5029860367700724,
+        0.4004672082940195,
+      ),
+      _Grad4(
+        0.7504883828755602,
+        0.4004672082940195,
+        -0.5029860367700724,
+        -0.15296486218853164,
+      ),
+      _Grad4(
+        0.6740059517812944,
+        0.3239847771997537,
+        -0.5794684678643381,
+        0.3239847771997537,
+      ),
+      _Grad4(
+        0.5794684678643381,
+        -0.3239847771997537,
+        -0.3239847771997537,
+        -0.6740059517812944,
+      ),
+      _Grad4(
+        0.5029860367700724,
+        -0.4004672082940195,
+        0.15296486218853164,
+        -0.7504883828755602,
+      ),
+      _Grad4(
+        0.5029860367700724,
+        0.15296486218853164,
+        -0.4004672082940195,
+        -0.7504883828755602,
+      ),
+      _Grad4(
+        0.4553054119602712,
+        0.08164729285680945,
+        0.08164729285680945,
+        -0.8828161875373585,
+      ),
+      _Grad4(
+        0.8828161875373585,
+        -0.08164729285680945,
+        -0.08164729285680945,
+        -0.4553054119602712,
+      ),
+      _Grad4(
+        0.7504883828755602,
+        -0.15296486218853164,
+        0.4004672082940195,
+        -0.5029860367700724,
+      ),
+      _Grad4(
+        0.7504883828755602,
+        0.4004672082940195,
+        -0.15296486218853164,
+        -0.5029860367700724,
+      ),
+      _Grad4(
+        0.6740059517812944,
+        0.3239847771997537,
+        0.3239847771997537,
+        -0.5794684678643381,
+      ),
+      _Grad4(
+        0.03381941603233842,
+        0.03381941603233842,
+        0.03381941603233842,
+        0.9982828964265062,
+      ),
+      _Grad4(
+        -0.044802370851755174,
+        -0.044802370851755174,
+        0.508629699630796,
+        0.8586508742123365,
+      ),
+      _Grad4(
+        -0.044802370851755174,
+        0.508629699630796,
+        -0.044802370851755174,
+        0.8586508742123365,
+      ),
+      _Grad4(
+        -0.12128480194602098,
+        0.4321472685365301,
+        0.4321472685365301,
+        0.7821684431180708,
+      ),
+      _Grad4(
+        0.508629699630796,
+        -0.044802370851755174,
+        -0.044802370851755174,
+        0.8586508742123365,
+      ),
+      _Grad4(
+        0.4321472685365301,
+        -0.12128480194602098,
+        0.4321472685365301,
+        0.7821684431180708,
+      ),
+      _Grad4(
+        0.4321472685365301,
+        0.4321472685365301,
+        -0.12128480194602098,
+        0.7821684431180708,
+      ),
+      _Grad4(
+        0.37968289875261624,
+        0.37968289875261624,
+        0.37968289875261624,
+        0.753341017856078,
+      ),
+      _Grad4(
+        0.03381941603233842,
+        0.03381941603233842,
+        0.9982828964265062,
+        0.03381941603233842,
+      ),
+      _Grad4(
+        -0.044802370851755174,
+        0.044802370851755174,
+        0.8586508742123365,
+        0.508629699630796,
+      ),
+      _Grad4(
+        -0.044802370851755174,
+        0.508629699630796,
+        0.8586508742123365,
+        -0.044802370851755174,
+      ),
+      _Grad4(
+        -0.12128480194602098,
+        0.4321472685365301,
+        0.7821684431180708,
+        0.4321472685365301,
+      ),
+      _Grad4(
+        0.508629699630796,
+        -0.044802370851755174,
+        0.8586508742123365,
+        -0.044802370851755174,
+      ),
+      _Grad4(
+        0.4321472685365301,
+        -0.12128480194602098,
+        0.7821684431180708,
+        0.4321472685365301,
+      ),
+      _Grad4(
+        0.4321472685365301,
+        0.4321472685365301,
+        0.7821684431180708,
+        -0.12128480194602098,
+      ),
+      _Grad4(
+        0.37968289875261624,
+        0.37968289875261624,
+        0.753341017856078,
+        0.37968289875261624,
+      ),
+      _Grad4(
+        0.03381941603233842,
+        0.9982828964265062,
+        0.03381941603233842,
+        0.03381941603233842,
+      ),
+      _Grad4(
+        -0.044802370851755174,
+        0.8586508742123365,
+        -0.044802370851755174,
+        0.508629699630796,
+      ),
+      _Grad4(
+        -0.044802370851755174,
+        0.8586508742123365,
+        0.508629699630796,
+        -0.044802370851755174,
+      ),
+      _Grad4(
+        -0.12128480194602098,
+        0.7821684431180708,
+        0.4321472685365301,
+        0.4321472685365301,
+      ),
+      _Grad4(
+        0.508629699630796,
+        0.8586508742123365,
+        -0.044802370851755174,
+        -0.044802370851755174,
+      ),
+      _Grad4(
+        0.4321472685365301,
+        0.7821684431180708,
+        -0.12128480194602098,
+        0.4321472685365301,
+      ),
+      _Grad4(
+        0.4321472685365301,
+        0.7821684431180708,
+        0.4321472685365301,
+        -0.12128480194602098,
+      ),
+      _Grad4(
+        0.37968289875261624,
+        0.753341017856078,
+        0.37968289875261624,
+        0.37968289875261624,
+      ),
+      _Grad4(
+        0.9982828964265062,
+        0.03381941603233842,
+        0.03381941603233842,
+        0.03381941603233842,
+      ),
+      _Grad4(
+        0.8586508742123365,
+        -0.044802370851755174,
+        -0.044802370851755174,
+        0.508629699630796,
+      ),
+      _Grad4(
+        0.8586508742123365,
+        -0.044802370851755174,
+        0.508629699630796,
+        -0.044802370851755174,
+      ),
+      _Grad4(
+        0.7821684431180708,
+        -0.12128480194602098,
+        0.4321472685365301,
+        0.4321472685365301,
+      ),
+      _Grad4(
+        0.8586508742123365,
+        0.508629699630796,
+        -0.044802370851755174,
+        -0.044802370851755174,
+      ),
+      _Grad4(
+        0.7821684431180708,
+        0.4321472685365301,
+        -0.12128480194602098,
+        0.4321472685365301,
+      ),
+      _Grad4(
+        0.7821684431180708,
+        0.4321472685365301,
+        0.4321472685365301,
+        -0.12128480194602098,
+      ),
+      _Grad4(
+        0.753341017856078,
+        0.37968289875261624,
+        0.37968289875261624,
+        0.37968289875261624,
+      ),
+    ];
+    final grad4Adjusted = [
+      for (final grad in grad4)
+        _Grad4(grad.dx / _kN4, grad.dy / _kN4, grad.dz / _kN4, grad.dw / _kN4),
+    ];
+    for (int i = 0; i < _kSize; i++) {
+      _gradients4d.add(grad4Adjusted[i % grad4.length]);
+    }
+  }
+}
+
+class _LatticePoint4D {
+  _LatticePoint4D(int xsv, int ysv, int zsv, int wsv)
+    : xsv = xsv + 409,
+      ysv = ysv + 409,
+      zsv = zsv + 409,
+      wsv = wsv + 409,
+      xsi = 0.2 - xsv,
+      ysi = 0.2 - ysv,
+      zsi = 0.2 - zsv,
+      wsi = 0.2 - wsv {
+    final ssv = (xsv + ysv + zsv + wsv) * 0.309016994374947;
+    dx = -xsv - ssv;
+    dy = -ysv - ssv;
+    dz = -zsv - ssv;
+    dw = -wsv - ssv;
+    ssiDelta = (0.8 - xsv - ysv - zsv - wsv) * 0.309016994374947;
+  }
+
+  final int xsv, ysv, zsv, wsv;
+  late final double dx, dy, dz, dw;
+  final double xsi, ysi, zsi, wsi;
+  late final double ssiDelta;
+}
