--- qaul_ui/packages/utils/lib/src/open_simplex_2.dart
@@ -1,5 +1,0 @@
-/// Abstract base class for OpenSimplex2 noise.
-abstract class OpenSimplex2 {
-  /// 4D noise, classic lattice orientation.
-  double noise4Classic(double x, double y, double z, double w);
-}
